requirements.txt
albumentations==1.4.21
pytorch-lightning
torchmetrics
wandb
loguru
segmentation-models-pytorch
numpy==1.24.4